Remove getHeaderTitle from hooks; added getHeaderTitle helper function
