+++ src/utils/helpers/index.ts
@@ -35,3 +35,15 @@ export const shortImageName = (imgName: string, length: number) => {
 
   return imgName;
 };
+
+export const getHeaderTitle = (pathName: string): string => {
+  const arr = pathName.split("/");
+
+  const pathNameArr = arr[arr.length - 1].split("-");
+
+  return pathNameArr
+    .map((s) => {
+      return s.charAt(0).toUpperCase() + s.slice(1);
+    })
+    .join(" ");
+};
